requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
